Pulled all the shared code into a wrapper.
